DEV: Ignore dropped Bookmark columns.
This is a partial revert of 099b679fc58b909d3800bc3f7a336982eedf4824.
`Bookmark#topic_id` and `Bookmark#reminder_type` was dropped in
b22450c7a8d378b32291743ca602cbc477dea106 so we need to continue ignoring
the dropped columns so as to ensure a seamless deploy. Otherwise,
ActiveRecord's schema cache will still contain references to
`Bookmark#topic_id` when the column is dropped in a post migration.