FIX: Emoji group not actually saving (#15602)
Follow up to 48f70dcd5f5e0c151342efd094f79cba8949872f. The group
_appeared_ to be saved in the UI until a refresh when it became
clear that the group wasn't actually sent to the DB. This is because
of the way the per-file data was being set with a computed property.
This commit fixes the computed property by changing it to a regular
function and also makes sure the name resetting after the first upload
in multiple uploads works too.